Move searchNavData to util dir
